Database added

+++ Pulsmesser/app/src/main/java/com/example/pulsmesser/Coordinate.java
@@ -4,10 +4,10 @@ import java.util.Date;
 
 public class Coordinate {
     public Date x;
-    public int y;
+    public float y;
 
     public Coordinate(String x, String y){
         this.x = new Date(Long.parseLong(x));
-        this.y = Integer.parseInt(y);
+        this.y = Float.parseFloat(y);
     }
 }
